requirements.txt
xlsxwriter
pandas
pydantic
google-genai
dotenv
openpyxl
fastapi
uvicorn